database + rela + fix

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -21,8 +21,8 @@ class CreateUsersTable extends Migration
             $table->string('address')->nullable();
             $table->date('birthday');
             $table->string('avatar')->default('default.png');
-            $table->float('salary');
-            $table->float('salary_ot');
+            $table->double('salary', 8, 2);
+            $table->double('salary_ot', 8, 2);
             $table->softDeletes();
             $table->rememberToken();
             $table->timestamps();
